Moved `convertToDate` and `convertToTime`.
Also introduced a `Holiday` model.
